+++ src/collaboration-service/controllers/session-controller.js
