+++ src/collaboration-service/models/session-model.js
